# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "먹어도 돼? API",
    "description" : "지병이 있는 사람이 관광지에서 외식할 때, 그 메뉴를 먹어도 되는지 /\n무엇을 빼고 먹어야 하는지 알려주는 서비스의 백엔드 API.\n\n## 인증\n\n자체 ID/PW + JWT. 로그인 응답의 `accessToken` 을 `Authorization: Bearer <토큰>` 으로 보낸다.\n\n- **필수**: `/api/me/**`, `/api/order-cards/**`, `POST /api/reports`\n- **선택**: `/api/restaurants/**` — 토큰이 없으면 목록만, 있으면 **사용자 기준 판정**이 함께 온다.\n  이때 응답의 `personalized` 로 어느 쪽인지 알 수 있다.\n- **불필요**: `/api/auth/**`, `GET /api/reports/options`\n\n## 판정값 (`seal`)\n\n색만으로 전달하지 않는다. 기호와 라벨을 항상 함께 노출할 것.\n\n| verdict | symbol | label |\n|---|---|---|\n| `RED` | ✕ | 알레르기 주의 |\n| `INK` | △ | 조절하면 가능 |\n| `OK` | ○ | 먹어도 돼요! |\n\n메뉴 판정은 **비관적**(하나라도 걸리면 위험), 식당 판정은 **낙관적**(안전한 메뉴가\n하나라도 있으면 갈 만함)이다. 방향이 반대인 것은 의도된 설계다.\n\n`seal` 이 `null` 인 경우가 있다. 아직 **분석되지 않은(미태깅) 메뉴/식당**이라는 뜻이며,\n**안전하다는 뜻이 아니다.** 프론트는 이 경우 판정 배지를 그리지 말고\n\"분석 전\" 으로 표시해야 한다.\n\n## 오류 형식\n\n모든 오류는 `{ \"code\": \"...\", \"message\": \"...\" }` 형태다.\n\n## 주의\n\n이 API 의 판정은 참고용이며 의학적 조언이 아니다. 응답의 `disclaimer` 문구를\n화면에 그대로 노출할 것.\n",
    "contact" : {
      "name" : "먹어도 돼? 백엔드"
    },
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "로컬"
  } ],
...
  "paths" : {
...
    "/api/reports" : {
      "post" : {
        "tags" : [ "5. 제보" ],
        "summary" : "제보 등록",
        "description" : "`feedback` 은 **정해진 문구 중에서** 고른 것만 받는다 (`GET /api/reports/options`).\n자유 문장은 `note` 에 넣을 것. flag 도출의 근거가 되는 값이라 자유 입력을 섞으면\n집계가 깨진다.\n\n`requests` 는 그때 실제로 사용한 요청 문구다 — 다른 사용자에게 주문 방법으로 보인다.",
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "등록됨",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReviewView"
                }
              }
            }
          },
          "401" : {
            "description" : "토큰이 없거나 만료됨",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReviewView"
                }
              }
            }
          },
          "404" : {
            "description" : "그런 식당이 없음",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReviewView"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/restaurants" : {
      "get" : {
        "tags" : [ "3. 식당" ],
        "summary" : "GPS 반경 검색",
        "description" : "현재 좌표에서 반경 안의 식당을 가까운 순으로 돌려준다.\n\n식당 판정은 **낙관적**이다 — 안전한 메뉴가 하나라도 있으면 `OK`,\n하나도 없고 조절 가능한 것이 있으면 `INK`, 전부 걸리면 `RED`.\n\n`seal` 이 `null` 이면 **아직 분석되지 않은 식당**이다. 안전하다는 뜻이 아니므로\n배지를 그리지 말 것.",
        "operationId" : "search",
        "parameters" : [ {
          "name" : "lat",
          "in" : "query",
          "description" : "위도",
          "required" : true,
          "schema" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 90,
            "minimum" : -90
          },
          "example" : 37.7519
        }, {
          "name" : "lng",
          "in" : "query",
          "description" : "경도",
          "required" : true,
          "schema" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 180,
            "minimum" : -180
          },
          "example" : 128.8761
        }, {
          "name" : "radius",
          "in" : "query",
          "description" : "반경(m). 비우면 서버 기본값",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          },
          "example" : 1000
        }, {
          "name" : "q",
          "in" : "query",
          "description" : "식당 이름 검색어",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "example" : "초당순두부"
        }, {
          "name" : "flags",
          "in" : "query",
          "description" : "제보에서 도출된 속성으로 거르기. 여러 개면 모두 만족하는 곳만",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "0부터 시작",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          },
          "example" : 0
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "한 쪽 개수. 0 이면 서버 기본값",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          },
          "example" : 20
        } ],
        "responses" : {
          "200" : {
            "description" : "검색됨. 결과가 없으면 items 가 빈 배열",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "좌표 범위를 벗어남",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/restaurants/{id}" : {
      "get" : {
        "tags" : [ "3. 식당" ],
        "summary" : "식당 상세",
        "description" : "메뉴별 판정까지 함께 온다. 메뉴 판정은 **비관적**이다 — 주재료 알레르겐이\n내 알레르기와 겹치면 `RED`, 양념에 미량으로 들어간 것(`hitTraceAllergens`)이나\n주의성분이 걸리면 `INK`.\n\n메뉴의 `tagStatus` 가 미태깅이면 `seal` 이 비어 있다. 그 메뉴는 판정하지 않은\n것이지 안전한 것이 아니다.\n\n좌표(`lat`,`lng`)를 함께 주면 거리·도보 시간도 채워 준다. 안 주면 `null`.",
        "operationId" : "detail",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "식당 id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          },
          "example" : 1
        }, {
          "name" : "lat",
          "in" : "query",
          "description" : "현재 위도. 주면 거리를 계산해 준다",
          "required" : false,
          "schema" : {
            "type" : "number",
            "format" : "double"
          },
          "example" : 37.7519
        }, {
          "name" : "lng",
          "in" : "query",
          "description" : "현재 경도",
          "required" : false,
          "schema" : {
            "type" : "number",
            "format" : "double"
          },
          "example" : 128.8761
        } ],
        "responses" : {
          "200" : {
            "description" : "조회됨",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestaurantDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "그런 식당이 없음",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestaurantDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/restaurants/{id}/reports" : {
      "get" : {
        "tags" : [ "5. 제보" ],
        "summary" : "식당의 제보 목록",
        "description" : "**비로그인도 볼 수 있다.** 토큰이 있으면 내가 쓴 제보에 `mine=true` 가 붙는다.\n\n`author` 는 `\"60대 · 당뇨\"` 처럼 파생한 값이다. 질환을 전부 나열하면 개인을\n좁힐 수 있어 하나만 노출한다. 나이·이름은 그대로 내보내지 않는다.\n\n`derivedFlags` 가 2건 이상 모여 식당에 실제로 붙은 속성이다.",
        "operationId" : "forRestaurant",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "식당 id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          },
          "example" : 1
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReviewListResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/me/reports" : {
      "get" : {
        "tags" : [ "5. 제보" ],
        "summary" : "내 제보 목록",
        "description" : "내가 쓴 제보 = 내 방문 기록. 최신순.",
        "operationId" : "mine",
        "responses" : {
          "401" : {
            "description" : "토큰이 없거나 만료됨",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ReviewView"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/me/reports/{id}" : {
      "delete" : {
        "tags" : [ "5. 제보" ],
        "summary" : "내 제보 삭제",
        "description" : "내 제보만 지울 수 있다. 지우면 그 식당의 flag 를 다시 계산한다.",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "제보 id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          },
          "example" : 1
        } ],
        "responses" : {
          "204" : {
            "description" : "지워짐. 본문 없음"
          },
          "401" : {
            "description" : "토큰이 없거나 만료됨"
          },
          "404" : {
            "description" : "없거나 내 제보가 아님"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "CreateRequest" : {
        "type" : "object",
        "description" : "제보 등록 요청",
        "properties" : {
          "restaurantId" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "어느 식당에 대한 제보인지",
            "example" : 1
          },
          "ok" : {
            "type" : "boolean",
            "description" : "요청이 받아들여졌는지 — 가게에 대한 평가",
            "example" : true
          },
          "note" : {
            "type" : "string",
            "description" : "자유 메모. 자유 문장은 여기에만 넣을 것",
            "maxLength" : 1000,
            "minLength" : 0
          },
          "requests" : {
            "type" : "array",
            "description" : "그때 실제로 사용한 요청 문구 — 다른 사용자에게 주문 방법으로 보인다",
            "example" : [ "국물은 따로 주세요" ],
            "items" : {
              "type" : "string"
            },
            "maxItems" : 20,
            "minItems" : 0
          },
          "feedback" : {
            "type" : "array",
            "description" : "GET /api/reports/options 의 문구 중에서 고른 것. flag 도출의 근거라\n**문구를 그대로** 보내야 한다. 철자가 다르면 집계에서 빠진다.",
            "example" : [ "덜짜게 해줌" ],
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          }
        },
        "required" : [ "ok", "restaurantId" ]
      },
      "ReviewView" : {
        "type" : "object",
        "description" : "제보 1건",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "author" : {
            "type" : "string",
            "description" : "작성자 표기. 나이·이름 대신 파생한 값이다",
            "example" : "60대 · 당뇨"
          },
          "mine" : {
            "type" : "boolean",
            "description" : "내가 쓴 제보인지. 비로그인이면 항상 false",
            "example" : false
          },
          "date" : {
            "type" : "string",
            "format" : "date",
            "description" : "방문한 날",
            "example" : "2026-09-11"
          },
          "ok" : {
            "type" : "boolean",
            "description" : "요청이 받아들여졌는지",
            "example" : true
          },
          "symbol" : {
            "type" : "string",
            "description" : "ok 에 대응하는 기호",
            "example" : "○"
          },
          "note" : {
            "type" : "string",
            "description" : "자유 메모"
          },
          "requests" : {
            "type" : "array",
            "description" : "그때 사용한 요청 문구",
            "items" : {
              "type" : "string"
            }
          },
          "feedback" : {
            "type" : "array",
            "description" : "고른 피드백 문구",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "RestaurantSummary" : {
        "type" : "object",
        "description" : "검색 결과의 식당 1건. 판정은 낙관적이다 — 안전한 메뉴가 하나라도 있으면 OK",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string",
            "example" : "초당할머니순두부"
          },
          "meta" : {
            "type" : "string",
            "description" : "주소·업종 등 한 줄 요약"
          },
          "lat" : {
            "type" : "number"
          },
          "lng" : {
            "type" : "number"
          },
          "distanceM" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "현재 좌표로부터의 직선 거리(m)",
            "example" : 320
          },
          "walkMinutes" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "도보 예상 시간(분)",
            "example" : 5
          },
          "flags" : {
            "type" : "array",
            "description" : "제보 2건 이상으로 도출된 속성",
            "example" : [ "덜짜게 해줌" ],
            "items" : {
              "type" : "string"
            }
          },
          "seal" : {
            "$ref" : "#/components/schemas/Seal",
            "description" : "비로그인이거나 메뉴가 전부 미태깅이면 null"
          },
          "summary" : {
            "type" : "string",
            "description" : "판정 근거 한 줄 요약"
          },
          "firstImage" : {
            "type" : "string",
            "description" : "대표 이미지 URL"
          }
        }
      },
      "Seal" : {
        "type" : "object",
        "description" : "판정 도장. **색만으로 전달하지 말 것** — symbol 과 label 을 항상 함께 노출한다\n(색각 이상 고려). 이 객체 자체가 null 이면 아직 분석되지 않았다는 뜻이며,\n안전하다는 뜻이 **아니다.**",
        "properties" : {
          "verdict" : {
            "type" : "string",
            "description" : "판정값",
            "enum" : [ "RED", "INK", "OK" ],
            "example" : "OK"
          },
          "symbol" : {
            "type" : "string",
            "description" : "RED=✕, INK=△, OK=○",
            "example" : "○"
          },
          "label" : {
            "type" : "string",
            "description" : "사람이 읽는 라벨",
            "example" : "먹어도 돼요!"
          }
        }
      },
      "SearchResponse" : {
        "type" : "object",
        "description" : "검색 결과",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 0
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 20
          },
          "totalCount" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "반경 안의 전체 건수",
            "example" : 37
          },
          "personalized" : {
            "type" : "boolean",
            "description" : "false 면 토큰이 없어 판정을 못 한 것이다. 이때 seal 은 전부 null",
            "example" : true
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RestaurantSummary"
            }
          },
          "disclaimer" : {
            "type" : "string",
            "description" : "화면에 그대로 노출해야 하는 고지 문구"
          }
        }
      },
      "MenuTagView" : {
        "type" : "object",
        "description" : "메뉴에 붙은 태그 1건 (알레르겐 또는 주의성분)",
        "properties" : {
          "value" : {
            "type" : "string",
            "description" : "성분 이름",
            "example" : "밀"
          },
          "amount" : {
            "type" : "string",
            "description" : "알레르겐의 함량 구분. `MAIN` 은 주재료라 ✕, `TRACE` 는 간장·된장 같은\n양념에 미량으로 들어간 것이라 △ 로 판정한다.",
            "enum" : [ "MAIN", "TRACE" ],
            "example" : "TRACE"
          },
          "source" : {
            "type" : "string",
            "description" : "이 태그가 어디서 왔는지 (LLM 태깅 / 영양성분 DB 등)",
            "example" : "LLM"
          },
          "estimated" : {
            "type" : "boolean",
            "description" : "true 면 추정값이다. 확정된 성분표가 아니다",
            "example" : true
          }
        }
      },
      "MenuView" : {
        "type" : "object",
        "description" : "메뉴 1건. 판정은 비관적이다 — 하나라도 걸리면 위험한 쪽으로 본다",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string",
            "example" : "초당순두부"
          },
          "price" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "가격. **항상 null 이다.** 어떤 공공 API 에도 가격이 없어 싣지 않는다\n(크롤링은 약관 위반, LLM 추정은 현장 피해). 화면에서 가격은 감출 것."
          },
          "representative" : {
            "type" : "boolean",
            "description" : "대표 메뉴 여부"
          },
          "tagStatus" : {
            "type" : "string",
            "description" : "태깅 상태. 미태깅이면 seal 이 비어 있다 — 안전한 것이 아니라 아직 분석 전이다",
            "example" : "TAGGED"
          },
          "seal" : {
            "$ref" : "#/components/schemas/Seal",
            "description" : "미태깅이거나 비로그인이면 null"
          },
          "tags" : {
            "type" : "array",
            "description" : "이 메뉴에 붙은 전체 태그",
            "items" : {
              "$ref" : "#/components/schemas/MenuTagView"
            }
          },
          "hitMainAllergens" : {
            "type" : "array",
            "description" : "내 알레르기와 겹친 **주재료** 알레르겐. 비어 있지 않으면 RED",
            "example" : [ "새우" ],
            "items" : {
              "type" : "string"
            }
          },
          "hitTraceAllergens" : {
            "type" : "array",
            "description" : "내 알레르기와 겹친 **양념 미량** 알레르겐. 이것만 있으면 INK",
            "example" : [ "밀" ],
            "items" : {
              "type" : "string"
            }
          },
          "hitCares" : {
            "type" : "array",
            "description" : "내 주의성분과 겹친 것",
            "example" : [ "나트륨" ],
            "items" : {
              "type" : "string"
            }
          },
          "detail" : {
            "type" : "string",
            "description" : "왜 이런 판정이 나왔는지 사람이 읽는 설명"
          },
          "suggestedRequests" : {
            "type" : "array",
            "description" : "이 메뉴에 쓸 만한 요청 문구. 주문요청카드의 기본값이 된다",
            "example" : [ "국물은 따로 주세요" ],
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "RestaurantDetail" : {
        "type" : "object",
        "description" : "식당 상세. 메뉴별 판정까지 포함한다",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string",
            "example" : "초당할머니순두부"
          },
          "meta" : {
            "type" : "string",
            "description" : "주소·업종 등 한 줄 요약"
          },
          "lat" : {
            "type" : "number"
          },
          "lng" : {
            "type" : "number"
          },
          "distanceM" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "요청에 lat,lng 를 주지 않으면 null"
          },
          "walkMinutes" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "요청에 lat,lng 를 주지 않으면 null"
          },
          "tel" : {
            "type" : "string"
          },
          "openTime" : {
            "type" : "string"
          },
          "restDate" : {
            "type" : "string"
          },
          "parking" : {
            "type" : "string"
          },
          "firstImage" : {
            "type" : "string"
          },
          "flags" : {
            "type" : "array",
            "description" : "제보 2건 이상으로 도출된 속성",
            "items" : {
              "type" : "string"
            }
          },
          "seal" : {
            "$ref" : "#/components/schemas/Seal",
            "description" : "식당 전체 판정 (낙관적). 미태깅이면 null"
          },
          "personalized" : {
            "type" : "boolean",
            "description" : "false 면 토큰이 없어 판정을 못 한 것"
          },
          "menus" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MenuView"
            }
          },
          "disclaimer" : {
            "type" : "string",
            "description" : "화면에 그대로 노출해야 하는 고지 문구"
          }
        }
      },
      "ReviewListResponse" : {
        "type" : "object",
        "description" : "식당의 제보 목록",
        "properties" : {
          "restaurantId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "count" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "제보 수",
            "example" : 3
          },
          "reviews" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewView"
            }
          },
          "derivedFlags" : {
            "type" : "array",
            "description" : "같은 피드백이 2건 이상 모여 이 식당에 실제로 붙은 속성. 검색 filter 에 쓴다",
            "example" : [ "덜짜게 해줌" ],
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "로그인(`POST /api/auth/login`) 또는 회원가입 응답의 `accessToken` 을 넣는다.\nSwagger UI 오른쪽 위 **Authorize** 에 토큰 값만 붙여 넣으면 된다 (`Bearer ` 접두사는 UI 가 붙인다).",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}